+++ src/main/java/net/elemental_wizards_rpg/effect/CleansingWaterEffect.java
@@ -4,13 +4,13 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
+import net.minecraft.particle.ParticleTypes;
 import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.Identifier;
+import net.minecraft.world.World;
+import net.more_rpg_classes.client.particle.MoreParticles;
 import net.more_rpg_classes.util.CustomMethods;
-import net.spell_engine.particle.ParticleHelper;
+import net.minecraft.server.world.ServerWorld;
 
-import static net.elemental_wizards_rpg.ElementalMod.MOD_ID;
-import static net.spell_engine.internals.SpellRegistry.getSpell;
 
 public class CleansingWaterEffect extends StatusEffect {
     public CleansingWaterEffect(StatusEffectCategory statusEffectCategory, int color) {
@@ -20,20 +20,36 @@ public class CleansingWaterEffect extends StatusEffect {
 
     @Override
     public void applyUpdateEffect(LivingEntity pLivingEntity, int pAmplifier) {
+        World world = pLivingEntity.getEntityWorld();
+
         if (pLivingEntity.getHealth() < pLivingEntity.getMaxHealth()) {
             pLivingEntity.heal(1.0F);
         }
         if(pLivingEntity.isOnFire()){
             pLivingEntity.playSound(SoundEvents.BLOCK_FIRE_EXTINGUISH,2,1);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[0].particles);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[1].particles);
+            if(world.isClient){
+                world.addParticle(MoreParticles.WATER_MIST,1,1,1,2,2,2);
+                world.addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
+            }else{
+                if (world instanceof ServerWorld serverWorld) {
+                    serverWorld.spawnParticles(MoreParticles.WATER_MIST,1,1,1,4,2,2,2,2);
+                    serverWorld.spawnParticles(ParticleTypes.LARGE_SMOKE,1,1,1,4,2,2,2,2);
+                }
+            }
             pLivingEntity.extinguish();
         }
         if(pLivingEntity.isInLava()){
-            pLivingEntity.removeStatusEffect(Effects.CLEANSING_WATER);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[0].particles);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[1].particles);
+            if(world.isClient){
+                world.addParticle(MoreParticles.WATER_MIST,1,1,1,2,2,2);
+                world.addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
+            }else{
+                if (world instanceof ServerWorld serverWorld) {
+                    serverWorld.spawnParticles(MoreParticles.WATER_MIST,1,1,1,4,2,2,2,2);
+                    serverWorld.spawnParticles(ParticleTypes.LARGE_SMOKE,1,1,1,4,2,2,2,2);
+                }
+            }
             pLivingEntity.playSound(SoundEvents.BLOCK_FIRE_EXTINGUISH,2,1);
+            pLivingEntity.removeStatusEffect(Effects.CLEANSING_WATER);
         }
     }