+++ src/main/java/net/elemental_wizards_rpg/effect/SoakedEffect.java
@@ -2,13 +2,13 @@ package net.elemental_wizards_rpg.effect;
 
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffectCategory;
+import net.minecraft.particle.ParticleTypes;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.Identifier;
-import net.spell_engine.particle.ParticleHelper;
+import net.minecraft.world.World;
+import net.more_rpg_classes.client.particle.MoreParticles;
 import net.spell_power.api.statuseffects.SpellVulnerabilityStatusEffect;
 
-import static net.elemental_wizards_rpg.ElementalMod.MOD_ID;
-import static net.spell_engine.internals.SpellRegistry.getSpell;
 
 public class SoakedEffect extends SpellVulnerabilityStatusEffect {
     public SoakedEffect(StatusEffectCategory statusEffectCategory, int color) {
@@ -17,20 +17,34 @@ public class SoakedEffect extends SpellVulnerabilityStatusEffect {
 
     @Override
     public void applyUpdateEffect(LivingEntity pLivingEntity, int pAmplifier) {
+        World world = pLivingEntity.getEntityWorld();
+
         if(pLivingEntity.isOnFire()){
-            pLivingEntity.removeStatusEffect(Effects.SOAKED);
+            if(world.isClient){
+                world.addParticle(MoreParticles.WATER_MIST,1,1,1,2,2,2);
+                world.addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
+            }else{
+                if (world instanceof ServerWorld serverWorld) {
+                    serverWorld.spawnParticles(MoreParticles.WATER_MIST,1,1,1,4,2,2,2,2);
+                    serverWorld.spawnParticles(ParticleTypes.LARGE_SMOKE,1,1,1,4,2,2,2,2);
+                }
+            }
             pLivingEntity.playSound(SoundEvents.BLOCK_FIRE_EXTINGUISH,2,1);
-            //pLivingEntity.getEntityWorld().addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[0].particles);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[1].particles);
             pLivingEntity.extinguish();
+            pLivingEntity.removeStatusEffect(Effects.SOAKED);
         }
         if(pLivingEntity.isInLava()){
-            pLivingEntity.removeStatusEffect(Effects.SOAKED);
+            if(world.isClient){
+                pLivingEntity.getEntityWorld().addParticle(MoreParticles.WATER_MIST,1,1,1,2,2,2);
+                pLivingEntity.getEntityWorld().addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
+            }else{
+                if (world instanceof ServerWorld serverWorld) {
+                    serverWorld.spawnParticles(MoreParticles.WATER_MIST,1,1,1,4,2,2,2,2);
+                    serverWorld.spawnParticles(ParticleTypes.LARGE_SMOKE,1,1,1,4,2,2,2,2);
+                }
+            }
             pLivingEntity.playSound(SoundEvents.BLOCK_FIRE_EXTINGUISH,2,1);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[0].particles);
-            ParticleHelper.sendBatches(pLivingEntity, getSpell(new Identifier(MOD_ID, "aqua_scald")).impact[1].particles);
-            //pLivingEntity.getEntityWorld().addParticle(ParticleTypes.LARGE_SMOKE,1,1,1,2,2,2);
+            pLivingEntity.removeStatusEffect(Effects.SOAKED);
         }
     }